requirements.txt
dhanhq
supabase
discord-webhook
pandas
numpy
python-dotenv
schedule
pytz
